fix: cast decimals to uint8 in morpho view

+++ contracts/views/MorphoAaveV2View.sol
@@ -16,7 +16,7 @@ contract MorphoAaveV2View is MorphoHelper {
     struct MarketInfo {
         address market; // aToken
         address underlying;
-        uint256 decimals;
+        uint8 decimals;
         uint256 p2pSupplyAmount;
         uint256 poolSupplyAmount;
         uint256 p2pBorrowAmount;
@@ -38,7 +38,7 @@ contract MorphoAaveV2View is MorphoHelper {
     struct UserBalance {
         address market;
         address underlying;
-        uint256 decimals;
+        uint8 decimals;
         uint256 userSupplyRate;
         uint256 userBorrowRate;
         uint256 supplyBalanceInP2P;
@@ -63,7 +63,7 @@ contract MorphoAaveV2View is MorphoHelper {
         return MarketInfo({
             market: _market,
             underlying: underlying,
-            decimals: IERC20(underlying).decimals(),
+            decimals: uint8(IERC20(underlying).decimals()),
             p2pSupplyAmount: p2pSupplyAmount,
             poolSupplyAmount: poolSupplyAmount,
             p2pBorrowAmount: p2pBorrowAmount,
@@ -118,7 +118,7 @@ contract MorphoAaveV2View is MorphoHelper {
             userInfo.userBalances[i] = UserBalance({
                 market: markets[i],
                 underlying: underlying,
-                decimals: IERC20(underlying).decimals(),
+                decimals: uint8(IERC20(underlying).decimals()),
                 userSupplyRate: IMorphoAaveV2Lens(MORPHO_AAVEV2_LENS_ADDR).getCurrentUserSupplyRatePerYear(markets[i], _usr),
                 userBorrowRate: IMorphoAaveV2Lens(MORPHO_AAVEV2_LENS_ADDR).getCurrentUserBorrowRatePerYear(markets[i], _usr),
                 supplyBalanceInP2P: supplyBalanceInP2P,
